+++ packages/elements/dynamic-table/src/styled/rankable/TableCell.js
@@ -1,0 +1,11 @@
+// @flow
+import styled, { css } from 'styled-components';
+import { TableBodyCell } from '../TableCell';
+
+const rankingStyles = css`
+  box-sizing: border-box;
+`;
+
+export const RankableTableBodyCell = styled(TableBodyCell)`
+  ${({ isRanking }) => isRanking && rankingStyles};
+`;